Move XNodeVisitor to sanity_utils
